+++ src/Resources/app/storefront/src/smart-payment-buttons/swag-paypal.smart-payment-buttons.js
@@ -211,7 +211,6 @@ export default class SwagPayPalSmartPaymentButtons extends SwagPaypalAbstractBut
                 this.options.createOrderUrl,
                 formData,
                 (responseText, request) => {
-                    PageLoadingIndicatorUtil.create();
                     if (request.status >= 400) {
                         reject(responseText);
                     }