+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - PPI-228 - Added color white as choice for ECS and SPB buttons
 - PPI-321 - Improve Zettle synchronisation behaviour
 - PPI-322 - Improve removal of PayPal from available payment methods if credentials are invalid
+- PPI-323 - Fix issues with loading spinner with Smart Payment Buttons
 
 # 3.3.1
 - PPI-316 - Fix issue when changing default language after plugin install